requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ML inference
torch==2.10.0
torchvision
efficientnet-pytorch==0.7.1
Pillow==10.3.0
numpy==1.26.4
onnxruntime==1.18.0   # optional: for ONNX deployment

# Groq
groq==0.9.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0

# Training (only needed for fine-tuning, not inference)
# scikit-learn==1.5.0
# matplotlib==3.9.0
# tqdm==4.66.4
# albumentations==1.4.8
